Deletion ob non-existent objects in database
git-svn-id: https://phpwebbuilder.svn.sourceforge.net/svnroot/phpwebbuilder/trunk@569 e3e4e2bd-1310-0410-aae6-9fb0b3a765dc

+++ Model/PersistentObject.class.php
@@ -161,6 +161,7 @@ class PersistentObject extends DescriptedObject {
 	}
 
 	function basicDelete() {
+		if (!$this->existsObject) return true;
 		$sql = 'DELETE FROM ' . $this->tableName() . ' WHERE id=' . $this->getId();
 		$db =& DB::Instance();
 		$can = TRUE;
